Version 1.1
+ added # (explicit string concatenation) operator.
- implicit string concatenation (without operators) now disabled by
default
it can be enabled in the SimpleContext (see
SimpleContext::implicitConcatenation() method)
* conditional and boolean logic was changed. Now string '0' (also empty
array) is considered as 'true'. NULL, false, 0, 0.0, '' are still
'false'
* & (or), | (and) and ^^ (exclusive or) are not boolean operators
anymore. Their result depend on the operands types:
A | B  equals to  A ? A : B
A & B  equals to  A ? B : A
A ^^ B equals to  !A ? B : (!B ? A : '')
* engine now has OR-chain processor which is used when construction A |
B | C ... is detected, it returns the first true operand, or the last
one if not a such.
+ several new optimizations.
- (X * 0) => 0 optimization was removed to handle right NAN and INF
values of the X expression